requirements.txt
python-dotenv
playwright
readability-lxml
lxml_html_clean
requests
